+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix: Fix `Triptych` component margin.
+
 ## [23.11.3] - 2018-11-14
 
 Fix: